requirements.txt
APScheduler==3.9.1.post1
asn1crypto==1.5.1
attrs==22.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
docker==6.0.1
executing==1.2.0
filelock==3.8.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==22.0
parsel==1.7.0
pg8000==1.29.4
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
scramp==1.4.4
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
Twisted==22.10.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
varname==0.10.0
w3lib==2.1.1
websocket-client==1.4.2
zope.interface==5.5.2
